+++ src/main.js
@@ -5,28 +5,50 @@ const util = require("util");
 const fs = require("fs");
 
 const defaults = {
-  optimizeAll: true,
+  optimizeAll: true, // optimize all images discovered in img tags
+
+  // Case insensitive. Only files whose extension exist in this array will be
+  // processed by the <img> tag (assuming `optimizeAll` above is true). Empty
+  // the array to allow all extensions to be processed. However, only jpegs and
+  // pngs are explicitly supported.
   imgTagExtensions: ["jpg", "jpeg", "png"],
+
+  // Same as the above, except that this array applies to the Image Component.
+  // If the images passed to your image component are unknown, it might be a
+  // good idea to populate this array.
   componentExtensions: [],
-  inlineBelow: 10000,
-  compressionLevel: 8,
-  quality: 70,
+
+  inlineBelow: 10000, // inline all images in img tags below 10kb
+
+  compressionLevel: 8, // png quality level
+
+  quality: 70, // jpeg/webp quality level
+
+  tagName: "Image", // default component name
+
+  sizes: [400, 800, 1200], // array of sizes for srcset in pixels
+
+  // array of screen size breakpoints at which sizes above will be applied
+  breakpoints: [375, 768, 1024],
+
+  outputDir: "g/",
+
+  placeholder: "trace", // or "blur",
+
+  // WebP options [sharp docs](https://sharp.pixelplumbing.com/en/stable/api-output/#webp)
   webpOptions: {
     quality: 75,
     lossless: false,
     force: true
   },
+
   webp: true,
-  tagName: "Image",
-  sizes: [400, 800, 1200],
-  breakpoints: [375, 768, 1024],
-  outputDir: "g/",
-  placeholder: "trace", // or "blur"
+
+  // Potrace options for SVG placeholder
   trace: {
     background: "#fff",
     color: "#002fa7",
-    threshold: 120,
-    size: 500
+    threshold: 120
   }
 };
 
@@ -401,6 +423,8 @@ async function replaceImages(content, options) {
   let ast;
   const imageNodes = [];
 
+  if (!content.includes("img") || !content.includes("Image")) return content;
+
   try {
     ast = svelte.parse(content);
   } catch {}
@@ -435,7 +459,7 @@ async function replaceImages(content, options) {
 }
 
 /**
- * @param {Partial<typeof defaults>} options 
+ * @param {Partial<typeof defaults>} options
  */
 function getPreprocessor(options = {}) {
   options = {
@@ -448,11 +472,10 @@ function getPreprocessor(options = {}) {
       code: await replaceImages(content, options)
     })
   };
-};
-
+}
 
 module.exports = {
   defaults,
   replaceImages,
   getPreprocessor
-}
+};