+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+
+
+## 0.1.0 - 2019-10-29 UTC
 ### Added
 - Extension filtering. `<img>` tags would incorrectly try to process files that
   were not processable, such as SVGs. Added an overridable list of file